requirements.txt
# common
Django==1.11.21
django-cors-headers==1.2.0
django-oauth-toolkit==0.10.0
djangorestframework==3.9.1
django-webpack-loader==0.6.0

# develop
flake8==3.7.8
